// src/generated/models/service-target-rule-model.ts
import type { ServiceTargetPolicy } from './service-target-policy-model';
import type { Workspace } from './workspace-model';

export const ServiceTargetRulePriorityKeyToLabel = {
  'Low': 'Low',
  'Medium': 'Medium',
  'High': 'High',
  'Critical': 'Critical'
} as const;
export type ServiceTargetRulePriorityKey = keyof typeof ServiceTargetRulePriorityKeyToLabel;

export interface ServiceTargetRule {
  /**
   * @displayName Service Target Rule ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Rule Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  ruleName: string;
  /**
   * @displayName Active
   * @validationRule Required for create/update operations
   */
  active: boolean;
  /**
   * @displayName Catalog Item Code
   */
  catalogItemCode?: string;
  /**
   * @displayName Field Operator Value Conditions
   * @validationRule Required for create/update operations
   */
  fieldOperatorValueConditions: string;
  /**
   * @displayName Impact
   */
  impact?: string;
  /**
   * @displayName Priority
   */
  priorityKey?: ServiceTargetRulePriorityKey;
  /**
   * @displayName Request Type Code
   */
  requestTypeCode?: string;
  /**
   * @displayName Service Code
   */
  serviceCode?: string;
  /**
   * @displayName Service Target Policy
   * @validationRule Required for create/update operations
   */
  serviceTargetPolicy: Pick<ServiceTargetPolicy, 'id' | 'name1'>;
  /**
   * @displayName Sort Order
   * @validationRule Required for create/update operations
   */
  sortOrder: number;
  /**
   * @displayName Urgency
   */
  urgency?: string;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _ServiceTargetRule = 'ServiceTargetRule' as const;